WIP dockchannel hid stuff
We need to upload the MTP firmware so I'm not sure I can get the keyboard will
working before being able to mount root. Another option is to embed the fw in
the kernel image (or u-boot?), but I need to double check the size to see if
that's feasible.