rtkit stuff
seems like I'm on the right track since RTKit responds as expected. At this
point I ran into some issues allocating DMA memory and realized that PAGE_SIZE
is 4K when it should be 16K. I think 16K pages might only be necessary for the
DART IOMMU which I don't think I'm using yet, so it might be possible to keep
going with 4K pages but it's probably best to try to get 16K pages working
instead. arrm64 with 16K pages is still experimental and there's definitely some
recent commits I need so I'm going to try to rebase this.