mk: disable BTI for arm64
The prebuilt core and alloc rlibs used by rustc are not
build with BTI so disable BTI in the rest of the kernel to
avoid linker errors. The proper solution is to either use
cargo and build-std and tell it to enable BTI or use custom
prebuilts of core and alloc built with BTI. I'm leaning
towards the former but it doesn't really matter for now.