Replace spinlock in apple_mbox.rs with split softc
Still trying to figure exactly how to track the state of the softc after the
split happens. Conceptually it's the same as the claimed_windows Vec in
kpi/bus.rs but it should be a smaller number which will certainly be known at
compile-time so even the minimal number of allocations may not end up being
necessary (especially if I can specify the number of contexts in the driver!
macro.